Move build_reference_value from utils.py to ComponentSpec class
